Code changes to test suite run and final cumulative report
