+++ src/main/java/com/freecrm/utils/Listeners/ExtentReporter.java
@@ -1,4 +1,4 @@
-package com.freecrm.utils.reports;
+package com.freecrm.utils.Listeners;
 
 import com.relevantcodes.extentreports.ExtentReports;
 import com.relevantcodes.extentreports.ExtentTest;
