+++ src/main/java/com/freecrm/utils/Listeners/ExtentReporterNG.java
@@ -11,13 +11,11 @@ import org.testng.IResultMap;
 import org.testng.ISuite;
 import org.testng.ISuiteResult;
 import org.testng.ITestContext;
-import org.testng.ITestNGListener;
 import org.testng.ITestResult;
 import org.testng.xml.XmlSuite;
 
 import com.relevantcodes.extentreports.ExtentReports;
 import com.relevantcodes.extentreports.ExtentTest;
-
 import com.relevantcodes.extentreports.LogStatus;
 
 public class ExtentReporterNG implements IReporter{
