+++ src/test/java/com/freecrm/test/login/FreeCRMLoginNew.java
@@ -8,30 +8,32 @@ import org.testng.Assert;
 import org.testng.ITestResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
+import com.freecrm.helper.base.BaseTest;
 import com.freecrm.po.FCHomePage;
-import com.freecrm.utils.ExcelDataInput;
+import com.freecrm.utils.ExcelUtil;
 
 
 
 public class FreeCRMLoginNew extends BaseTest 
 {	
-	@BeforeMethod(alwaysRun=true)
+	@BeforeMethod
 	void initialize(Method method) {
 		beforeMethod(method);
 		System.out.println("Driver Initialization Started...");		
 	}
 	
-	/*@DataProvider(name="getTestData")
+	@DataProvider(name="getTestData")
 	public Object[][] getAllTestData()
 	{
 		return ExcelUtil.getAllTestData("testdata", "Sheet1");
-	}*/
+	}
 	
-	@ExcelDataInput(filename="testdata.xlsx" , sheetname="Sheet1")
-	@Test(description="Free CRM Login Test" , groups= {"Author:Vineet"})
-	public void FreeCRMLoginTest(HashMap<String, String> testdata)
+	//@ExcelDataInput(filename="testdata.xlsx" , sheetname="Sheet1")
+	@Test(description="Free CRM Login Test" ,dataProvider="getTestData", groups= {"Author:Vineet"})
+	public void FreeCRMLoginTestNew(HashMap<String, String> testdata)
 	{	
 	
 		super.initializeSetUp();
@@ -53,7 +55,7 @@ public class FreeCRMLoginNew extends BaseTest
 		Reporter().log(Reporter.get().getStatus(), "Step 4 :: Switch to Main panel");
 		
 		Assert.assertEquals(homepage.verifyCRMPROLogoText(), false);
-		Reporter().log(Reporter.get().getStatus(), "Step 5 :: verifyCRMPROLogoText");
+		Reporter().log(Reporter.get().getStatus(), "Step 5 :: verify CRMPRO Logo Text");
 		
 	}
